feat(w-reader): 优化自动阅读功能并修复相关问题
-修复编辑器消息提示窗口宽度和高度设置异常问题
- 新增章节自动切换功能
- 支持自动阅读时间小数点设置
- 优化 EPUB 文件目录读取逻辑
- 调整图片预览初始化逻辑，适应窗口大小

+++ src/main/java/com/wei/wreader/pojo/Settings.java
@@ -68,7 +68,7 @@ public class Settings {
     /**
      * 自动阅读时间(秒)
      */
-    private int autoReadTime;
+    private float autoReadTime;
     /**
      * 本地加载时是否显示图片
      */
@@ -151,11 +151,11 @@ public class Settings {
         this.charset = charset;
     }
 
-    public int getAutoReadTime() {
+    public float getAutoReadTime() {
         return autoReadTime;
     }
 
-    public void setAutoReadTime(int autoReadTime) {
+    public void setAutoReadTime(float autoReadTime) {
         this.autoReadTime = autoReadTime;
     }
